lal_demodulate:  Bug fix for rounding negative frequencies.

+++ gstlal-calibration/gst/lal/gstlal_demodulate.h
@@ -79,7 +79,7 @@ struct _GSTLALDemodulate {
 	gboolean need_discont;
 
 	/* properties */
-	int line_frequency; /* in centihertz */
+	gint64 line_frequency; /* in uHz */
 	double prefactor_real;
 	double prefactor_imag;
 };
